Beginning moving hwm to class and QtCharts.
