+++ html/hwm_map.html
@@ -52,7 +52,7 @@
     <script src="https://maps.googleapis.com/maps/api/js?v=3.exp&sensor=false"></script>
     <script src="/rsc/js/hwm_map.js"></script>
   </head>
-  <body onload="initializeHWM()">
+  <body onload="initializeHWM()" style="overflow: hidden">
     <div id="map_canvas" style="width: 100%"></div>
     <div id="legend"></div>
   </body>
