+++ html/timeseries_maps.html
@@ -35,7 +35,7 @@
     <script src="https://maps.googleapis.com/maps/api/js?v=3.exp&sensor=false"></script>
     <script src="/rsc/js/timeseries_maps.js"></script>
   </head>
-  <body onload="initializeTimeseries()">
+  <body onload="initializeTimeseries()" style="overflow: hidden">
       <div style="height:100%; width: 100%; display: table">
           <div style="display: table-row">
             <div id="map_canvas" style="width: 100%; display: table-cell;"></div>
