+++ html/usgs_maps.html
@@ -50,7 +50,7 @@
     <script src="https://maps.googleapis.com/maps/api/js?v=3.exp&sensor=false"></script>
     <script src="/rsc/js/usgs_maps.js"></script>
   </head>
-  <body onload="initialize()">
+  <body onload="initialize()" style="overflow: hidden">
       <div style="height:100%; width: 100%; display: table">
           <div style="display: table-row">
             <div id="map_canvas" style="width: 50%; display: table-cell;"></div>
