+++ src/hwm.cpp
@@ -1,0 +1,7 @@
+#include "hwm.h"
+
+hwm::hwm(QObject *parent) : QObject(parent)
+{
+
+}
+
