+++ src/hwm_data.cpp
@@ -21,7 +21,7 @@
 //
 //-----------------------------------------------------------------------//
         
-#include <hwm.h>
+#include <hwm_orig.h>
 
 //-------------------------------------------//
 //A function that reads a CSV high water
